chore(validator): unify remote config validators codebase
