+++ agent-control/src/agent_control.rs
@@ -6,7 +6,6 @@ pub(super) mod event_handler;
 pub use agent_control::*;
 #[allow(clippy::module_inception)]
 mod agent_control;
-pub mod config_validator;
 pub mod http_server;
 pub mod pid_cache;
 pub mod run;
