+++ agent-control/src/opamp/remote_config/validators.rs
@@ -1,6 +1,10 @@
 use super::RemoteConfig;
-use crate::agent_control::config::AgentTypeFQN;
+use crate::agent_control::config::{AgentControlDynamicConfig, AgentTypeFQN};
+use crate::agent_type::agent_type_registry::AgentRepositoryError;
 use std::fmt::Display;
+use thiserror::Error;
+
+pub mod dynamic_config;
 pub mod regexes;
 pub mod signature;
 
@@ -12,6 +16,20 @@ pub trait RemoteConfigValidator {
         agent_type_fqn: &AgentTypeFQN,
         remote_config: &RemoteConfig,
     ) -> Result<(), Self::Err>;
+}
+
+#[derive(Error, Debug)]
+pub enum DynamicConfigValidatorError {
+    #[error("validating dynamic config`{0}`")]
+    AgentRepositoryError(#[from] AgentRepositoryError),
+}
+
+/// Represents a validator for dynamic config
+pub trait DynamicConfigValidator {
+    fn validate(
+        &self,
+        dynamic_config: &AgentControlDynamicConfig,
+    ) -> Result<(), DynamicConfigValidatorError>;
 }
 
 #[cfg(test)]
@@ -30,6 +48,17 @@ pub mod tests {
                 agent_type_fqn: &AgentTypeFQN,
                 remote_config: &RemoteConfig,
             ) -> Result<(), <Self as RemoteConfigValidator>::Err>;
+        }
+    }
+
+    mock! {
+        pub DynamicConfigValidatorMock {}
+
+        impl DynamicConfigValidator for DynamicConfigValidatorMock {
+            fn validate(
+                &self,
+                dynamic_config: &AgentControlDynamicConfig,
+            ) -> Result<(), DynamicConfigValidatorError>;
         }
     }
 }
